+++ src/mips/psyqo/src/gpu.cpp
@@ -27,7 +27,6 @@ SOFTWARE.
 #include "psyqo/gpu.hh"
 
 #include <EASTL/atomic.h>
-#include <EASTL/fixed_list.h>
 #include <EASTL/functional.h>
 
 #include "common/hardware/counters.h"
@@ -147,18 +146,6 @@ void psyqo::GPU::initialize(const psyqo::GPU::Configuration &config) {
     dicr |= 0x040000;
     Hardware::CPU::DICR = dicr;
 }
-
-namespace {
-struct Timer {
-    eastl::function<void(uint32_t)> callback;
-    uint32_t deadline;
-    uint32_t period;
-    int32_t pausedRemaining;
-    bool periodic;
-    bool paused = false;
-};
-eastl::fixed_list<Timer, 32> s_timers;
-}  // namespace
 
 void psyqo::GPU::flip() {
     do {
@@ -430,17 +417,17 @@ bool psyqo::GPU::isChainTransferred() const {
 }
 
 uintptr_t psyqo::GPU::armTimer(uint32_t deadline, eastl::function<void(uint32_t)> &&callback) {
-    s_timers.emplace_back(eastl::move(callback), deadline, 0, 0, false);
-    return reinterpret_cast<uintptr_t>(&s_timers.back());
+    m_timers.emplace_back(eastl::move(callback), deadline, 0, 0, false);
+    return reinterpret_cast<uintptr_t>(&m_timers.back());
 }
 
 uintptr_t psyqo::GPU::armPeriodicTimer(uint32_t interval, eastl::function<void(uint32_t)> &&callback) {
-    s_timers.emplace_back(eastl::move(callback), m_currentTime + interval, interval, 0, true);
-    return reinterpret_cast<uintptr_t>(&s_timers.back());
+    m_timers.emplace_back(eastl::move(callback), m_currentTime + interval, interval, 0, true);
+    return reinterpret_cast<uintptr_t>(&m_timers.back());
 }
 
 void psyqo::GPU::changeTimerPeriod(uintptr_t id, uint32_t period, bool reset) {
-    for (auto &timer : s_timers) {
+    for (auto &timer : m_timers) {
         if (reinterpret_cast<uintptr_t>(&timer) != id) continue;
         if (timer.period == period) continue;
         if (!timer.periodic) continue;
@@ -456,7 +443,7 @@ void psyqo::GPU::changeTimerPeriod(uintptr_t id, uint32_t period, bool reset) {
 }
 
 void psyqo::GPU::pauseTimer(uintptr_t id) {
-    for (auto &timer : s_timers) {
+    for (auto &timer : m_timers) {
         if (reinterpret_cast<uintptr_t>(&timer) != id) continue;
         if (timer.paused) return;
         timer.paused = true;
@@ -466,7 +453,7 @@ void psyqo::GPU::pauseTimer(uintptr_t id) {
 }
 
 void psyqo::GPU::resumeTimer(uintptr_t id) {
-    for (auto &timer : s_timers) {
+    for (auto &timer : m_timers) {
         if (reinterpret_cast<uintptr_t>(&timer) != id) continue;
         if (!timer.paused) return;
         timer.paused = false;
@@ -476,9 +463,9 @@ void psyqo::GPU::resumeTimer(uintptr_t id) {
 }
 
 void psyqo::GPU::cancelTimer(uintptr_t id) {
-    for (auto it = s_timers.begin(); it != s_timers.end(); ++it) {
+    for (auto it = m_timers.begin(); it != m_timers.end(); ++it) {
         if (reinterpret_cast<uintptr_t>(&*it) != id) continue;
-        s_timers.erase(it);
+        m_timers.erase(it);
         return;
     }
 }
@@ -493,7 +480,7 @@ void psyqo::GPU::pumpCallbacks() {
     bool done = false;
     while (!done) {
         done = true;
-        for (auto it = s_timers.begin(); it != s_timers.end(); it++) {
+        for (auto it = m_timers.begin(); it != m_timers.end(); it++) {
             auto &timer = *it;
             if (timer.paused) continue;
             if ((int32_t)(timer.deadline - currentTime) <= 0) {
@@ -502,7 +489,7 @@ void psyqo::GPU::pumpCallbacks() {
                 }
                 timer.callback(currentTime);
                 if (!timer.periodic) {
-                    s_timers.erase(it);
+                    m_timers.erase(it);
                 }
                 done = false;
                 break;
